refactor(onboarding): extract OptionsBubbleCta to share options CTA logic
Pulls the identical configureContentViews and setOnOptionClicked
implementations out of DaxTryASearchBrandDesignUpdateBubbleCta and
DaxVisitSiteOptionsBrandDesignUpdateBubbleCta into a shared
OptionsBubbleCta base so both subclasses reduce to their differing
constructor values.
