+++ app/src/main/java/com/duckduckgo/app/cta/ui/DaxVisitSiteOptionsBrandDesignUpdateBubbleCta.kt
@@ -16,70 +16,24 @@
 
 package com.duckduckgo.app.cta.ui
 
-import android.view.View
 import com.duckduckgo.app.browser.R
 import com.duckduckgo.app.cta.model.CtaId
 import com.duckduckgo.app.global.install.AppInstallStore
 import com.duckduckgo.app.onboarding.store.OnboardingStore
-import com.duckduckgo.app.onboarding.ui.view.OnboardingSelectionButton
-import com.duckduckgo.app.pixels.AppPixelName
 import com.duckduckgo.app.statistics.pixels.Pixel
-import com.duckduckgo.common.ui.view.gone
-import com.duckduckgo.common.ui.view.show
 
 data class DaxVisitSiteOptionsBrandDesignUpdateBubbleCta(
     override val onboardingStore: OnboardingStore,
     override val appInstallStore: AppInstallStore,
     override val isLightTheme: Boolean,
-) : DaxBubbleCta.BrandDesignUpdateBubbleCta(
+) : OptionsBubbleCta(
     ctaId = CtaId.DAX_INTRO_VISIT_SITE,
     title = R.string.onboardingSitesDaxDialogTitle,
     description = R.string.onboardingSitesDaxDialogDescription,
     options = onboardingStore.getSitesOptions(),
     backgroundRes = R.drawable.bg_onboarding_site_options,
-    shownPixel = AppPixelName.ONBOARDING_DAX_CTA_SHOWN,
-    okPixel = AppPixelName.ONBOARDING_DAX_CTA_OK_BUTTON,
     ctaPixelParam = Pixel.PixelValues.DAX_INITIAL_VISIT_SITE_CTA,
     onboardingStore = onboardingStore,
     appInstallStore = appInstallStore,
     isLightTheme = isLightTheme,
-) {
-    override val activeIncludeId: Int = R.id.optionsContent
-
-    override fun configureContentViews(view: View) {
-        val optionViews: List<OnboardingSelectionButton> = listOf(
-            view.findViewById(R.id.brandDesignOption1),
-            view.findViewById(R.id.brandDesignOption2),
-            view.findViewById(R.id.brandDesignOption3),
-            view.findViewById(R.id.brandDesignOption4),
-        )
-        options?.let {
-            optionViews.forEachIndexed { index, buttonView ->
-                buttonView.show()
-                if (it.size > index) {
-                    it[index].setOptionView(buttonView)
-                } else {
-                    buttonView.gone()
-                }
-            }
-        }
-    }
-
-    override fun setOnOptionClicked(
-        onboardingExperimentEnabled: Boolean,
-        configuration: DaxBubbleCta?,
-        onOptionClicked: (DaxDialogIntroOption, index: Int?) -> Unit,
-    ) {
-        options?.forEachIndexed { index, option ->
-            val optionViewId = when (index) {
-                0 -> R.id.brandDesignOption1
-                1 -> R.id.brandDesignOption2
-                2 -> R.id.brandDesignOption3
-                else -> R.id.brandDesignOption4
-            }
-            ctaView?.findViewById<OnboardingSelectionButton>(optionViewId)?.setOnClickListener {
-                onOptionClicked.invoke(option, index)
-            }
-        }
-    }
-}
+)
